refactor: centralize login validation
