+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Example script demonstrating AuthorService usage.
 - Documentation for pagination and asynchronous usage examples.
 - List of stable public API exports.
+- Base class to centralize login validation for authenticated services.
 ### Fixed
 - Avoided ``httpx`` deprecation warning when posting raw bytes or text.
